setting proper id data type

+++ src/main/java/com/travelapp/travelapp/dto/collagepost/CollageDTOPost.java
@@ -3,7 +3,7 @@ package com.travelapp.travelapp.dto.collagepost;
 import java.util.List;
 
 public record CollageDTOPost(
-        int userId,
+        Long userId,
         List<Integer> pictures,
         String description
 ) {
